[ME-1808] Service Daemon Compatibility with OpenBSD
